rename create -> lens_edit
