+++ apps/web/index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
+    <title>Spyglass Search</title>
     <meta charset="utf-8" />
     <link data-trunk rel="css" href="/public/main.css"/>
     <link data-trunk rel="copy-dir" href="/public/icons"/>
